refactor: wrap main routes in main module for coverage
